+++ cmd/sidecar/worker.go
@@ -395,89 +395,6 @@ func (worker *inboundWorker) prepFunctionFiles(ctx context.Context, ir *function
 	return statusCode, nil
 }
 
-func getReferencedFile(ctx context.Context, flowToken, flowAddr, namespace string, path string) ([]byte, int, error) {
-	addr := fmt.Sprintf("http://%v/api/v2/namespaces/%v/files/%v", flowAddr, namespace, path)
-	var d []byte
-
-	resp, err := doRequest(ctx, http.MethodGet, flowToken, addr, nil)
-	if resp.StatusCode == http.StatusNotFound {
-		// some very special magic
-		return nil, resp.StatusCode, &RessourceNotFoundError{
-			Key:   path,
-			Scope: "file",
-		}
-	}
-	if err != nil {
-		return nil, resp.StatusCode, fmt.Errorf("failed to do request: %w", err)
-	}
-	statusCode, err := handleResponse(resp, func(resp *http.Response) (int, error) {
-		var respData decodedFilesResponse
-		if resp.Body == nil {
-			return http.StatusInternalServerError, fmt.Errorf("unexpected failure: response body is nil")
-		}
-		decoder := json.NewDecoder(resp.Body)
-		if err := decoder.Decode(&respData); err != nil {
-			target := &json.UnmarshalTypeError{}
-			if errors.As(err, &target) {
-				return http.StatusBadRequest, fmt.Errorf("invalid response format: %w", err)
-			}
-			return http.StatusInternalServerError, fmt.Errorf("failed to decode response: %w", err)
-		}
-
-		if len(respData.Data.Data) > 0 {
-			d, err = base64.StdEncoding.DecodeString(respData.Data.Data)
-			if err != nil {
-				return http.StatusInternalServerError, fmt.Errorf("failed to decode base64 data: %w", err)
-			}
-		}
-		return http.StatusOK, nil
-	})
-	if err != nil {
-		return nil, statusCode, err
-	}
-
-	return d, statusCode, nil
-}
-
-func getVariables(ctx context.Context, flowToken, addr string) (*variablesResponse, int, error) {
-	resp, err := doRequest(ctx, http.MethodGet, flowToken, addr, nil)
-	if err != nil {
-		return nil, resp.StatusCode, err
-	}
-	variables := variablesResponse{}
-
-	statusCode, err := handleResponse(resp, func(resp *http.Response) (int, error) {
-		decoder := json.NewDecoder(resp.Body)
-		if err = decoder.Decode(&variables); err != nil {
-			return resp.StatusCode, fmt.Errorf("failed to decode response body: %w", err)
-		}
-		return http.StatusOK, nil
-	})
-	if err != nil {
-		return nil, statusCode, err
-	}
-
-	return &variables, resp.StatusCode, nil
-}
-
-func getInstanceVariables(ctx context.Context, flowToken string, flowAddr string, ir *functionRequest) (*variablesResponse, int, error) {
-	addr := fmt.Sprintf("http://%v/api/v2/namespaces/%v/variables?instanceId=%v", flowAddr, ir.namespace, ir.instanceId)
-
-	return getVariables(ctx, flowToken, addr)
-}
-
-func getNamespaceVariables(ctx context.Context, flowToken string, flowAddr string, ir *functionRequest) (*variablesResponse, int, error) {
-	addr := fmt.Sprintf("http://%v/api/v2/namespaces/%v/variables", flowAddr, ir.namespace)
-
-	return getVariables(ctx, flowToken, addr)
-}
-
-func getWorkflowVariables(ctx context.Context, flowToken string, flowAddr string, ir *functionRequest) (*variablesResponse, int, error) {
-	addr := fmt.Sprintf("http://%v/api/v2/namespaces/%v/variables?workflowPath=%v", flowAddr, ir.namespace, ir.workflowPath)
-
-	return getVariables(ctx, flowToken, addr)
-}
-
 func fetchFunctionFiles(ctx context.Context, flowToken string, flowAddr string, ir *functionRequest, fileWriter func(context.Context, *functionRequest, *functionFiles, *io.PipeReader) error) (int, error) {
 	namespaceVariables, statusCode, err := getNamespaceVariables(ctx, flowToken, flowAddr, ir)
 	if err != nil {
